Conslidated multiple js file and css to client folder
