+++ index.js
@@ -1,34 +1,34 @@
-const { app, BrowserWindow } = require('electron');
+const {app, BrowserWindow} = require('electron');
 
 const createWindow = () => {
-    const win = new BrowserWindow({
-        width: 400,
-        minWidth: 400,
-        height: 752,
-        minHeight: 752,
-        webPreferences: {
-            nodeIntegration: true,
-            contextIsolation: false
-        },
-        icon: 'assets/logo-1000.png',
-    })
+	const win = new BrowserWindow({
+		width: 400,
+		minWidth: 400,
+		height: 752,
+		minHeight: 752,
+		webPreferences: {
+			nodeIntegration: true,
+			contextIsolation: false
+		},
+		icon: 'assets/logo-1000.png',
+	})
 
-    win.resizable = true;
-    win.setAlwaysOnTop(true);
+	win.resizable = true;
+	win.setAlwaysOnTop(true);
 
-    win.loadFile('index.html')
+	win.loadFile('client/index.html')
 }
 
 app.setName("Trader Bruh");
 
 app.whenReady().then(() => {
-    createWindow()
+	createWindow()
 
-    app.on('activate', () => {
-        if (BrowserWindow.getAllWindows().length === 0) createWindow()
-    })
+	app.on('activate', () => {
+		if (BrowserWindow.getAllWindows().length === 0) createWindow()
+	})
 })
 
 app.on('window-all-closed', () => {
-    if (process.platform !== 'darwin') app.quit()
+	if (process.platform !== 'darwin') app.quit()
 })
