--- scripts/deps.js
@@ -1,1 +1,0 @@
-const $ = require('jquery');
