--- scripts/index.js
@@ -1,13 +1,0 @@
-var db = {};
-
-document.addEventListener('DOMContentLoaded', () => {
-    const modules = {
-        PositionSizeBuilder: PositionSizeBuilder($)
-    };
-
-    Object.values(modules).forEach(module => {
-        if (typeof module.init === 'function') {
-            module.init();
-        }
-    });
-});
